Add files via upload
Here are the slides of my talk at the symposium Continuous Norming: Recent Advancements in Research and Application, at the International Test Commission Conference.  

Abstract: To classify subjects’ performance on relevant clinical or educational measures, such as neuropsychological tests or language tests, psychologists and educators need reference values or norms. Norms are important because they allow practitioners to compare subjects’ performance on a test with the reference population determined by the purpose of the test (e.g., diagnosis of dementia versus job selection). Based on this comparison, psychologists and educators can make important decisions about individuals, such as assigning a patient to a treatment. For this reason, norms must be precise, that is, not strongly affected by sampling error in the sample on which the norms are based. In this talk, it will be shown how to minimize sampling error and maximize the precision of the norms in three steps. First, norms should be obtained by regressing the test score on predictors deemed relevant for the norming (e.g., age and sex) since this approach is more efficient than the traditional approach of splitting the sample into subgroups based on demographic factors and deriving norms per subgroup. Second, the precision of the norms can be maximized by carefully selecting the composition of the sample of the normative study (e.g., which age groups to include) for a given total sample size. Third, this sample size can be computed such that a prespecified power for classification, or a prespecified precision of estimation, is obtained. These three steps will be illustrated under two scenarios: norming of one test with multiple linear regression and norming of several tests with multivariate multiple linear regression.